[OB-3917] fix: change all int_64t exposed script parameters to int_32t
This is because our current version of quickjs does not support BigInt
(64 bits ints), which led to an error if using BigInt in a script on
Web, or an unsupported conversion warning if using a Number.

Now that we expose 32 bits ints, the web scripts no longer show an
unsupported conversion warning when using a Number.

+++ Library/src/Multiplayer/Script/ComponentBinding/ReflectionSpaceComponentScriptInterface.h
@@ -37,7 +37,7 @@ public:
     DECLARE_SCRIPT_PROPERTY(Vector3, Position);
     DECLARE_SCRIPT_PROPERTY(Vector3, Scale);
 
-    DECLARE_SCRIPT_PROPERTY(int64_t, ReflectionShape);
+    DECLARE_SCRIPT_PROPERTY(int32_t, ReflectionShape);
 };
 
 } // namespace csp::multiplayer
